1. cant delete id 0 2. relational built version 1 many to many

+++ src/main/java/com/example/students/system/controller/RestSysController.java
@@ -5,6 +5,8 @@ import com.example.students.system.dao.TeacherDao;
 import com.example.students.system.domain.Student;
 import com.example.students.system.dto.StudentDto;
 
+import com.example.students.system.exception.StudentException;
+import com.example.students.system.service.StudentEntityToDtoServiceImpl;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import org.modelmapper.ModelMapper;
@@ -16,23 +18,23 @@ import org.springframework.web.bind.annotation.RestController;
 
 import javax.validation.Valid;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @RestController
-@RequestMapping("/api")
+@RequestMapping("/api/student")
 @Api
 public class RestSysController {
     @Autowired
     StudentDao studentDao;
 
     @Autowired
-    StudentDto studentDto;
-
-    @Autowired
     TeacherDao teacherDao;
 
     @Autowired
-    ModelMapper modelMapper;
+    StudentEntityToDtoServiceImpl studentService;
 
+//    @Autowired
+//    StudentDto studentDto;
 //    @ApiOperation(value = "gets a single Student")
 //    @GetMapping(value = "/student/{id}")
 //    public ResponseEntity<Student> getStudent(@PathVariable("id") int id) {
@@ -46,53 +48,60 @@ public class RestSysController {
 //    }
     @ApiOperation(value = "gets a single Student")
     @GetMapping(value = "/student/{id}")
-    public ResponseEntity<StudentDto> getStudent(@PathVariable("id") int id) {
-        //String stu_id = String.valueOf(id);
-        List<Student> students = studentDao.findById(id);
+    public ResponseEntity<StudentDto> getStudent(@PathVariable("id") long id) {
+        Student student = studentDao.findOneById(id);
+        if (student==null)
+            throw new StudentException("User not found");
 
-        if (students.size() == 0) {
-            System.out.println("not found");
-        }
-        studentDto = modelMapper.map(students.get(0), StudentDto.class);
-
-        return new ResponseEntity<StudentDto>(studentDto, HttpStatus.OK);
+        return new ResponseEntity<StudentDto>(studentService.convertSingleObject(student), HttpStatus.OK);
     }
+
 
     @ApiOperation(value = "gets all Student")
     @GetMapping(value = "/student/")
-    public ResponseEntity<List<Student>> getStudentPagenation() {
-
-        List<Student> students = studentDao.findAll();
-        if (students.isEmpty()) {
-            //throw new UserNotFoundException(messages.getMessage("USER_NOT_FOUND"));
-        }
-        return new ResponseEntity<List<Student>>( students, HttpStatus.OK);
-
+    public ResponseEntity<List<StudentDto>> getStudentPagenation() {
+        List<StudentDto> studentsDto = studentService.convertAll(studentDao.findAll());
+        if(studentsDto.size()==0)
+            throw new StudentException("User not found");
+        else
+            return new ResponseEntity<List<StudentDto>>(studentsDto , HttpStatus.OK);
     }
 
     @ApiOperation(value = "insert a single getStudent")
     @PostMapping(value = "/student/")
-    public Student insertStudent(@Valid @RequestBody Student student) {
-        return studentDao.save(student);
+    public ResponseEntity<StudentDto> insertStudent(@Valid @RequestBody StudentDto studentDto) {
+        if(studentDao.findOneById((int)studentDto.getId())!=null)
+            throw new StudentException("User already exist");
+        Student student = new Student();
+        student.setFirstname(studentDto.getFirstName());
+        student.setLastname(studentDto.getLastName());
+        student.setId(studentDto.getId());
+        System.out.println("Id: "+ student.getId());
+        studentDao.save(student);
+        return new ResponseEntity<StudentDto>(studentDto, HttpStatus.OK);
     }
 
     @ApiOperation(value = "update a single getStudent")
     @PutMapping(value = "/student/")
-    public Student updateStudent(@Valid @RequestBody Student student) {
-        return studentDao.saveAndFlush(student);
+    public StudentDto updateStudent(@Valid @RequestBody StudentDto studentDto) {
+        Student student = studentDao.findOneById((long)studentDto.getId());
+        if (student==null)
+            throw new StudentException("User not found");
+        student.setFirstname(studentDto.getFirstName());
+        student.setLastname(studentDto.getLastName());
+        studentDao.saveAndFlush(student);
+        return studentDto;
     }
 
     @ApiOperation(value = "delete a single getStudent")
     @DeleteMapping(value = "/student/{id}")
-    public void deleteStudent(@PathVariable("id") int id) {
-        try{
-            studentDao.deleteById(id);
-            //return "delete successfully";
-        }catch (Exception e){
-
-        }
-        System.out.print("delete fail");
+    public ResponseEntity<String> deleteStudent(@PathVariable("id") long id) {
+        Student student = studentDao.findOneById(id);
+        if(student==null)
+            throw new StudentException("User not found");
+        studentDao.deleteById(id);
         //return "failed";
+        return new ResponseEntity<String>("Delete successfully id: "+id, HttpStatus.OK);
     }
